chore(engine): day 4 — move shell AST helpers to @node9/policy-engine
Pure mvdan-sh wrappers (normalizeCommandForPolicy, detectDangerousShellExec)
move into packages/policy-engine/src/shell/. Proxy re-exports for
back-compat. Each module gets its own cached parser instance until
analyzeShellCommand follows in day 8.
